@@ -9,7 +9,7 @@
     "type": "project",
     "version": "2.21.0",
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "^8.0",
         "sentry/sdk": "^3.1"
     },
     "homepage": "https://etymology.dsantini.it",
