@@ -13,12 +13,7 @@ use \App\Result\JSONRemoteQueryResult;
 
 class JSONCurlQuery extends CurlQuery implements JSONQuery
 {
-    /**
-     * @param string|null $result
-     * @param array $curlInfo
-     * @return QueryResult
-     */
-    protected function getResultFromCurlData($result, $curlInfo): QueryResult
+    protected function getResultFromCurlData(?string $result, array $curlInfo): QueryResult
     {
         return new JSONRemoteQueryResult($result, $curlInfo);
     }
